reports/requirements.txt
aws-cdk-lib==2.174.0
cdk-nag>=2.14.11
constructs>=10.0.0,<11.0.0
pre-commit
